+++ packages/linux_x86_64/cp311/exts/wandelbots.omni/wandelbots/omni/router/v2/base.py
@@ -47,7 +47,7 @@ API_DESCRIPTION = (
 omniservice_app = FastAPI(
     title=API_TITLE,
     description=API_DESCRIPTION,
-    version="2.34.3",
+    version="2.34.5",
     docs_url=None,
     redoc_url=None,
 )