+++ packages/linux_x86_64/cp311/exts/wandelbots.omni/wandelbots/omni/ui/tool/__init__.py
@@ -1,10 +1,16 @@
-from .nova_tcp import (
+from wandelbots.omni.ui.tool.nova_tcp import (
     register_tcp_from_isaac_to_nova_menu,
     register_tcp_from_nova_to_isaac_menu,
 )
-from .collision_setup.collision_setup_window import register_collision_setup_window
-from .action_planner.action_planner_window import register_action_planner_window
-from .ghost_teaching.ghost_teaching_tool_bar import register_ghost_teaching_tool_bar
+from wandelbots.omni.ui.tool.collision_setup.collision_setup_window import (
+    register_collision_setup_window,
+)
+from wandelbots.omni.ui.tool.action_planner.action_planner_window import (
+    register_action_planner_window,
+)
+from wandelbots.omni.ui.tool.ghost_teaching.ghost_teaching_tool_bar import (
+    register_ghost_teaching_tool_bar,
+)
 
 
 def register_tools():